requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arxiv==2.2.0
bibtexparser==1.4.3
certifi==2025.1.31
charset-normalizer==3.4.1
diskcache==5.6.3
distro==1.9.0
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.3.2
gitignore-parser==0.1.11
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
llama-cpp-python==0.3.8
loguru==0.7.3
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
openai==1.73.0
packaging==24.2
pillow==11.2.1
pip==25.0.1
pydantic==2.11.3
pydantic-core==2.33.1
pyparsing==3.2.3
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
pyzotero==1.6.11
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
setuptools==65.5.0
sgmllib3k==1.0.0
sniffio==1.3.1
sympy==1.13.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.48.3
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uv==0.6.14
beautifulsoup4>=4.12.0
